requirements.txt
streamlit
genai
python-dotenv
os
time
json
logging